+++ NEWS.rst
@@ -4,7 +4,7 @@ News
 1.3 (unreleased)
 ----------------
 * Detect redundant 'if' conditions without 'else' blocks.
-* Add whitelist for ``string.Formatter`` (#183).
+* Add whitelist for ``string.Formatter`` (Joseph Bylund, #183).
 
 
 1.2 (2019-11-22)